Tweak default settings of optimalGif for Capture's purpose
We found that the defaults here were speeding up the videos so the gifs
were significantly shorter than the original recordings. Set it to 15 fps
and 100 / 15 = 0.066 for delay between frames.

A less lazy person than me might've changed this to instead use `createGIFfromURL`
but we can cross that bridge if we want the user to have control over the
output of the GIF later.